Create a separate controller for files.
Basically just to break up the giant app file. The actual file methods
should also be refactored as they are all huge individually at the
moment.
