+++ config.ru
@@ -8,4 +8,4 @@ end
 
 require_relative "app/app"
 
-run Tint::App.new
+run Tint::App
